Add AR status and last editor to list view
