+++ src/views/water/abstraction-reform/edit.html
@@ -3,7 +3,7 @@
   {{>element-phase-tag}}
   {{>element-main-nav}}
 
-  <a class="link-back" href="/start">Back</a>
+  <a class="link-back" href="/admin/abstraction-reform/licence/{{ documentId }}">Back</a>
 
   <div class="grid-row">
     <div class="column-full">
